Make responses nullable for the time being
