+++ mod.ts
@@ -7,6 +7,7 @@ export * from "./drivers/driver.ts";
 export * from "./drivers/core_adapter.ts";
 export * from "./registration.ts";
 export * from "./route_decorators.ts";
+export * from "./post_decorator.ts";
 export * from "./model_decorators.ts";
 export * from "./model_builders.ts";
 export * from "./context.ts";
